Moved home fragment test where it belongs
